pyproject.toml
[project]
name = "wiki-creator"
version = "0.1.0"
description = "Bootstraps an LLM-maintained wiki into any target project"
requires-python = ">=3.12"
dependencies = [
    "jinja2>=3.1",
    "pyyaml>=6.0",
    "click>=8.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-snapshot>=0.9",
    "pytest-cov>=5.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-snapshot>=0.9",
    "pytest-cov>=5.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q"

[tool.coverage.run]
source = ["tools", "scripts"]
omit = ["tests/*"]

[tool.hatch.build.targets.wheel]
packages = ["tools", "scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
